[Refactor] useFetchColumns
